move ResultManagement specifications
